Improve the AArch32 PrintDisassembler class.
* Add Decode*At() methods.
* Modify the Disassemble*Buffer() methods to guard against buffer
overflow.
* Change the buffer pointers to byte pointers (as for AArch64).

+++ examples/aarch32/disasm-a32.cc
@@ -296,12 +296,12 @@ int main(int argc, char** argv) {
     if ((func_addr & 1) == 1) {
       func_addr &= ~1;
       dis.JumpToPc(func_addr);
-      dis.DisassembleT32Buffer(reinterpret_cast<uint16_t*>(
+      dis.DisassembleT32Buffer(reinterpret_cast<uint8_t*>(
                                    base_addr + symbol.GetMemoryAddress()),
                                func_size);
     } else {
       dis.JumpToPc(func_addr);
-      dis.DisassembleA32Buffer(reinterpret_cast<uint32_t*>(
+      dis.DisassembleA32Buffer(reinterpret_cast<uint8_t*>(
                                    base_addr + symbol.GetMemoryAddress()),
                                func_size);
     }
